chainlit/requirements.txt
pytesseract==0.3.10
pillow==10.1.0
chainlit==0.7.2
langchain==0.0.315
pymupdf==1.23.5
tiktoken== 0.5.1
openai==0.28.1
pypdf2==3.0.1
faiss-cpu==1.7.4
pandas==2.1.3
openpyxl==3.1.2